Update to addToCart functionality

+++ database/migrations/2021_05_08_100237_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->decimal('price', 8, 2, true);
+            $table->float('price', 8, 2, true);
             $table->timestamps();
         });
     }
